Test code cleanup.
