+++ src/MphRead/Formats/Effects.cs
@@ -142,44 +142,44 @@ namespace MphRead.Effects
 
         protected void FxFunc05(IReadOnlyList<int> param, TimeValues times, ref Vector3 vec)
         {
-            vec.X = Fixed.ToFloat(Test.GetRandomInt1(4096));
-            vec.Y = Fixed.ToFloat(Test.GetRandomInt1(4096));
-            vec.Z = Fixed.ToFloat(Test.GetRandomInt1(4096));
+            vec.X = Fixed.ToFloat(Rng.GetRandomInt1(4096));
+            vec.Y = Fixed.ToFloat(Rng.GetRandomInt1(4096));
+            vec.Z = Fixed.ToFloat(Rng.GetRandomInt1(4096));
         }
 
         protected void FxFunc06(IReadOnlyList<int> param, TimeValues times, ref Vector3 vec)
         {
-            vec.X = Fixed.ToFloat(Test.GetRandomInt1(4096));
+            vec.X = Fixed.ToFloat(Rng.GetRandomInt1(4096));
             vec.Y = 0;
-            vec.Z = Fixed.ToFloat(Test.GetRandomInt1(4096));
+            vec.Z = Fixed.ToFloat(Rng.GetRandomInt1(4096));
         }
 
         protected void FxFunc07(IReadOnlyList<int> param, TimeValues times, ref Vector3 vec)
         {
-            vec.X = Fixed.ToFloat(Test.GetRandomInt1(4096));
+            vec.X = Fixed.ToFloat(Rng.GetRandomInt1(4096));
             vec.Y = 1;
-            vec.Z = Fixed.ToFloat(Test.GetRandomInt1(4096));
+            vec.Z = Fixed.ToFloat(Rng.GetRandomInt1(4096));
         }
 
         protected void FxFunc08(IReadOnlyList<int> param, TimeValues times, ref Vector3 vec)
         {
-            vec.X = Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f;
-            vec.Y = Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f;
-            vec.Z = Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f;
+            vec.X = Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f;
+            vec.Y = Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f;
+            vec.Z = Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f;
         }
 
         protected void FxFunc09(IReadOnlyList<int> param, TimeValues times, ref Vector3 vec)
         {
-            vec.X = Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f;
+            vec.X = Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f;
             vec.Y = 0;
-            vec.Z = Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f;
+            vec.Z = Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f;
         }
 
         protected void FxFunc10(IReadOnlyList<int> param, TimeValues times, ref Vector3 vec)
         {
-            vec.X = Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f;
+            vec.X = Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f;
             vec.Y = 1;
-            vec.Z = Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f;
+            vec.Z = Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f;
         }
 
         protected abstract void FxFunc11(IReadOnlyList<int> param, TimeValues times, ref Vector3 vec);
@@ -224,7 +224,7 @@ namespace MphRead.Effects
         {
             float value1 = InvokeFloatFunc(Funcs[(uint)param[0]], times);
             float value2 = InvokeFloatFunc(Funcs[(uint)param[1]], times);
-            float angle = MathHelper.DegreesToRadians((Test.GetRandomInt1(0xFFFF) >> 4) * (360 / 4096f));
+            float angle = MathHelper.DegreesToRadians((Rng.GetRandomInt1(0xFFFF) >> 4) * (360 / 4096f));
             vec.X = MathF.Sin(angle) * value1;
             vec.Y = value2;
             vec.Z = MathF.Cos(angle) * value1;
@@ -234,9 +234,9 @@ namespace MphRead.Effects
         {
             float value1 = InvokeFloatFunc(Funcs[(uint)param[0]], times);
             float value2 = InvokeFloatFunc(Funcs[(uint)param[1]], times);
-            vec.X = (Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f) * value1;
+            vec.X = (Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f) * value1;
             vec.Y = 0;
-            vec.Z = (Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f) * value2;
+            vec.Z = (Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f) * value2;
         }
 
         protected void FxFunc17(IReadOnlyList<int> param, TimeValues times, ref Vector3 vec)
@@ -369,18 +369,18 @@ namespace MphRead.Effects
 
         protected float FxFunc43(IReadOnlyList<int> param, TimeValues times)
         {
-            return Fixed.ToFloat(Test.GetRandomInt1(4096));
+            return Fixed.ToFloat(Rng.GetRandomInt1(4096));
         }
 
         protected float FxFunc44(IReadOnlyList<int> param, TimeValues times)
         {
-            return Fixed.ToFloat(Test.GetRandomInt1(4096)) - 0.5f;
+            return Fixed.ToFloat(Rng.GetRandomInt1(4096)) - 0.5f;
         }
 
         // get random angle [0-360) in fx32
         protected float FxFunc45(IReadOnlyList<int> param, TimeValues times)
         {
-            return Fixed.ToFloat(Test.GetRandomInt1(0x168000));
+            return Fixed.ToFloat(Rng.GetRandomInt1(0x168000));
         }
 
         protected float FxFunc46(IReadOnlyList<int> param, TimeValues times)
