+++ src/MphRead/Renderer.cs
@@ -786,8 +786,8 @@ namespace MphRead
 
         public void OnRenderFrame(double frameTime)
         {
-            uint rng1 = Test.Rng1;
-            uint rng2 = Test.Rng2;
+            uint rng1 = Rng.Rng1;
+            uint rng2 = Rng.Rng2;
             if (_recording || Debugger.IsAttached)
             {
                 _frameTime = 1 / 60f; // todo: FPS stuff
@@ -814,8 +814,8 @@ namespace MphRead
             RenderScene();
             if (_frameAdvanceOn && !_advanceOneFrame)
             {
-                Test.SetRng1(rng1);
-                Test.SetRng2(rng2);
+                Rng.SetRng1(rng1);
+                Rng.SetRng2(rng2);
             }
             if (!_frameAdvanceOn || _advanceOneFrame)
             {
