+++ src/MphRead/Testing/TestWeapons.cs
@@ -1,0 +1,1208 @@
+using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Runtime.InteropServices;
+
+namespace MphRead.Testing
+{
+    public static class TestWeapons
+    {
+        private static string EnumToString<T>(T value) where T : Enum
+        {
+            string name = typeof(T).Name;
+            return $"{name}.{value}".Replace(", ", $" | {name}.");
+        }
+
+        public static void TestWeaponInfo()
+        {
+            IReadOnlyList<WeaponInfo> weapons1P = Weapons.Weapons1P;
+            IReadOnlyList<WeaponInfo> weaponsMP = Weapons.WeaponsMP;
+            IReadOnlyList<WeaponInfo> enemies = Weapons.EnemyWeapons;
+            IReadOnlyList<WeaponInfo> platforms = Weapons.PlatformWeapons;
+            IReadOnlyList<WeaponInfo> ricochets = Weapons.Ricochets;
+
+            foreach (WeaponInfo weapon in enemies)
+            {
+                var ids = new HashSet<ushort>();
+                ids.Add(weapon.DrawFuncIds[0]);
+                ids.Add(weapon.DrawFuncIds[1]);
+                if (ids.Contains(6) || ids.Contains(12))
+                {
+                    Debugger.Break();
+                }
+            }
+
+            Nop();
+
+            for (int i = 0; i < 9; i++)
+            {
+                WeaponInfo singleNormal = weapons1P[i];
+                WeaponInfo multiNormal = weaponsMP[i];
+                WeaponInfo singleAffinity = weapons1P[i + 9];
+                WeaponInfo multiAffinity = weaponsMP[i + 9];
+
+                Console.WriteLine(singleNormal.Name);
+                Console.WriteLine($"1P Nrm: {singleNormal.UnchargedSpread} / {singleNormal.MinChargeSpread} / {singleNormal.ChargedSpread}");
+                Console.WriteLine($"MP Nrm: {multiNormal.UnchargedSpread} / {multiNormal.MinChargeSpread} / {multiNormal.ChargedSpread}");
+                Console.WriteLine($"1P Aff: {singleAffinity.UnchargedSpread} / {singleAffinity.MinChargeSpread} / {singleAffinity.ChargedSpread}");
+                Console.WriteLine($"MP Aff: {multiAffinity.UnchargedSpread} / {multiAffinity.MinChargeSpread} / {multiAffinity.ChargedSpread}");
+                Console.WriteLine();
+
+                //int bit = 9;
+                //uint mask = (uint)Math.Pow(2, bit);
+                //bool snTest = (singleNormal.Flags & mask) != 0;
+                //bool mnTest = (multiNormal.Flags & mask) != 0;
+                //bool saTest = (singleAffinity.Flags & mask) != 0;
+                //bool maTest = (multiAffinity.Flags & mask) != 0;
+                //var results = new HashSet<bool>();
+                //results.Add(snTest);
+                //results.Add(mnTest);
+                //results.Add(saTest);
+                //results.Add(maTest);
+                //if (results.Count != 1)
+                //{
+                //    Debugger.Break();
+                //}
+                //Console.WriteLine(singleNormal.Name);
+                //Console.WriteLine($"1P Nrm: {snTest}");
+                //Console.WriteLine($"MP Nrm: {mnTest}");
+                //Console.WriteLine($"1P Aff: {saTest}");
+                //Console.WriteLine($"MP Aff: {maTest}");
+                //Console.WriteLine();
+            }
+
+            Nop();
+
+            //for (int i = 0; i < ricochets.Count; i++)
+            //{
+            //    WeaponInfo weapon = ricochets[i];
+            //    if (weapon.Description != "")
+            //    {
+            //        Console.WriteLine(weapon.Description);
+            //    }
+            //    else
+            //    {
+            //        Console.WriteLine($"{i} / {weapon.Weapon} / {weapon.WeaponType}");
+            //    }
+            //    Console.WriteLine($"{weapon.DrawFuncIds[0]} {weapon.DrawFuncIds[1]}");
+            //    Console.WriteLine();
+            //}
+
+            //static void WriteStuff(WeaponInfo w, bool aff, bool mp)
+            //{
+            //    Console.WriteLine($"{w.Name}" + (aff ? " (AF)" : "(NM)") + (mp ? " (MP)" : " (1P)"));
+            //    byte afl1 = w.Afflictions[0];
+            //    string str1 = "";
+            //    if ((afl1 & 1) != 0)
+            //    {
+            //        str1 += "Frz ";
+            //    }
+            //    if ((afl1 & 2) != 0)
+            //    {
+            //        str1 += "Dis ";
+            //    }
+            //    if ((afl1 & 4) != 0)
+            //    {
+            //        str1 += "Brn ";
+            //    }
+            //    Debug.Assert((afl1 & 0xF8) == 0);
+            //    byte afl2 = w.Afflictions[1];
+            //    string str2 = "";
+            //    if ((afl2 & 1) != 0)
+            //    {
+            //        str2 += "Frz ";
+            //    }
+            //    if ((afl2 & 2) != 0)
+            //    {
+            //        str2 += "Dis ";
+            //    }
+            //    if ((afl2 & 4) != 0)
+            //    {
+            //        str2 += "Brn ";
+            //    }
+            //    str1 = str1.Trim();
+            //    str2 = str2.Trim();
+            //    Debug.Assert((afl2 & 0xF8) == 0);
+            //    Console.WriteLine($"{str1} / {str2}");
+            //}
+
+            //for (int i = 0; i < 9; i++)
+            //{
+            //    WriteStuff(weapons1P[i], false, false);
+            //    WriteStuff(weapons1P[i + 9], true, false);
+            //    WriteStuff(weaponsMP[i], false, true);
+            //    WriteStuff(weaponsMP[i + 9], true, true);
+            //    Console.WriteLine();
+            //}
+        }
+
+        public static IReadOnlyList<RawWeaponInfo> ParseWeaponInfo(int count, byte[] array)
+        {
+            int size = Marshal.SizeOf<RawWeaponInfo>();
+            Debug.Assert(size == 0xF0);
+            var results = new List<RawWeaponInfo>();
+            var bytes = new ReadOnlySpan<byte>(array);
+            Debug.Assert(bytes.Length == count * size);
+            for (int i = 0; i < count; i++)
+            {
+                int start = i * size;
+                int end = start + size;
+                results.Add(Read.ReadStruct<RawWeaponInfo>(bytes[start..end]));
+                //results.Add(MemoryMarshal.Read<WeaponInfo>(bytes[start..end]));
+            }
+            return results;
+        }
+
+        public static void DumpWeaponInfo(RawWeaponInfo weapon)
+        {
+            Console.WriteLine($@"            new WeaponInfo(
+                description: """",
+                weapon: BeamType.{weapon.Weapon},
+                weaponType: BeamType.{weapon.WeaponType},
+                drawFuncIds: new byte[] {{ {weapon.DrawFuncIds[0]}, {weapon.DrawFuncIds[1]} }},
+                colors: new ushort[] {{ {weapon.Colors[0]}, {weapon.Colors[1]} }},
+                flags: {EnumToString(weapon.Flags)},
+                splashDamage: {weapon.SplashDamage},
+                minChargeSplashDamage: {weapon.MinChargeSplashDamage},
+                chargedSplashDamage: {weapon.ChargedSplashDamage},
+                splashDmgTypes: new byte[] {{ {weapon.SplashDmgTypes[0]}, {weapon.SplashDmgTypes[1]} }},
+                shotCooldown: {weapon.ShotCooldown},
+                shotCooldownRelated: {weapon.ShotCooldownRelated},
+                ammoType: {weapon.AmmoType},
+                beamTypes: new byte[] {{ {weapon.BeamTypes[0]}, {weapon.BeamTypes[1]} }},
+                muzzleEffects: new byte[] {{ {weapon.MuzzleEffects[0]}, {weapon.MuzzleEffects[1]} }},
+                dmgDirTypes: new byte[] {{ {weapon.DmgDirTypes[0]}, {weapon.DmgDirTypes[1]} }},
+                field1D: new byte[] {{ {weapon.Field1D[0]}, {weapon.Field1D[1]} }},
+                afflictions: new Affliction[] {{ {EnumToString(weapon.Afflictions[0])}, {EnumToString(weapon.Afflictions[1])} }},
+                field21: {weapon.Field21},
+                minCharge: {weapon.MinCharge},
+                fullCharge: {weapon.FullCharge},
+                ammoCost: {weapon.AmmoCost},
+                minChargeCost: {weapon.MinChargeCost},
+                chargeCost: {weapon.ChargeCost},
+                unchargedDamage: {weapon.UnchargedDamage},
+                minChargeDamage: {weapon.MinChargeDamage},
+                chargedDamage: {weapon.ChargedDamage},
+                headshotDamage: {weapon.HeadshotDamage},
+                minChargeHeadshotDamage: {weapon.MinChargeHeadshotDamage},
+                chargedHeadshotDamage: {weapon.ChargedHeadshotDamage},
+                unchargedLifespan: {weapon.UnchargedLifespan},
+                minChargeLifespan: {weapon.MinChargeLifespan},
+                chargedLifespan: {weapon.ChargedLifespan},
+                speedDecay: new ushort[] {{ {weapon.SpeedDecay[0]}, {weapon.SpeedDecay[1]} }},
+                field42: {weapon.Field42},
+                speedInterp: new ushort[] {{ {weapon.SpeedInterp[0]}, {weapon.SpeedInterp[1]} }},
+                field48: {weapon.Field48},
+                field4C: {weapon.Field4C},
+                field50: {weapon.Field50},
+                field54: {weapon.Field54},
+                field58: {weapon.Field58},
+                field5C: {weapon.Field5C},
+                field60: {weapon.Field60},
+                unchargedSpeed: {weapon.UnchargedSpeed},
+                minChargeSpeed: {weapon.MinChargeSpeed},
+                chargedSpeed: {weapon.ChargedSpeed},
+                unchargedFinalSpeed: {weapon.UnchargedFinalSpeed},
+                minChargeFinalSpeed: {weapon.MinChargeFinalSpeed},
+                chargedFinalSpeed: {weapon.ChargedFinalSpeed},
+                unchargedGravity: {weapon.UnchargedGravity},
+                minChargeGravity: {weapon.MinChargeGravity},
+                chargedGravity: {weapon.ChargedGravity},
+                unchargedHoming: {weapon.UnchargedHoming},
+                minChargeHoming: {weapon.MinChargeHoming},
+                chargedHoming: {weapon.ChargedHoming},
+                homingRange: {weapon.HomingRange},
+                homingTolerance: {weapon.HomingTolerance},
+                unchargedScale: {weapon.UnchargedScale},
+                minChargeScale: {weapon.MinChargeScale},
+                chargedScale: {weapon.ChargedScale},
+                unchargedDistance: {weapon.UnchargedDistance},
+                minChargeDistance: {weapon.MinChargeDistance},
+                chargedDistance: {weapon.ChargedDistance},
+                unchargedSpread: {weapon.UnchargedSpread},
+                minChargeSpread: {weapon.MinChargeSpread},
+                chargedSpread: {weapon.ChargedSpread},
+                fieldC0: {weapon.FieldC0},
+                fieldC4: {weapon.FieldC4},
+                fieldC8: {weapon.FieldC8},
+                fieldCC: {weapon.FieldCC},
+                fieldD0: {weapon.FieldD0},
+                fieldD4: {weapon.FieldD4},
+                ricochetWeapon: new uint[] {{ 0x{weapon.RicochetWeaponPtr[0]:X1}, 0x{weapon.RicochetWeaponPtr[1]:X1} }},
+                projectileCount: {weapon.ProjectileCount},
+                minChargedProjectileCount: {weapon.MinChargedProjectileCount},
+                chargeProjectileCount: {weapon.ChargeProjectileCount},
+                smokeStart: {weapon.SmokeStart},
+                smokeMinimum: {weapon.SmokeMinimum},
+                smokeDrain: {weapon.SmokeDrain},
+                smokeShotAmount: {weapon.SmokeShotAmount},
+                smokeChargeAmount: {weapon.SmokeChargeAmount}
+            ),");
+        }
+
+        private static IReadOnlyList<RawWeaponInfo> GetRicochets()
+        {
+            return ParseWeaponInfo(count: 4, new byte[]
+            {
+                // 0 - 0x20BDAB0 - Judicator
+                0x05, 0x05, 0x03, 0x03, 0x94, 0x7E, 0x94, 0x7E, 0x02, 0x12, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x14, 0x14, 0x00, 0x0B, 0x0B, 0xFF, 0xFF, 0x03, 0x03, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x00, 0x18, 0x00,
+                0x18, 0x00, 0x18, 0x00, 0x18, 0x00, 0x18, 0x00, 0x0F, 0x00, 0x0F, 0x00, 0x0F, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00,
+                0x33, 0x03, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x90, 0x01, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x32, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00,
+                // 1 - 0x20CA130
+                0x06, 0x0A, 0x04, 0x04, 0x3F, 0x02, 0x3F, 0x02, 0x02, 0x01, 0x00, 0x40, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x02, 0x02, 0x14, 0x14, 0x00, 0x09, 0x09, 0xFF, 0xFF, 0x03, 0x03, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00, 0x05, 0x00,
+                0x05, 0x00, 0x05, 0x00, 0x05, 0x00, 0x05, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x02, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x34, 0xFF, 0xFF, 0xFF,
+                0x34, 0xFF, 0xFF, 0xFF, 0x34, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00,
+                0x00, 0x30, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x40, 0xA0, 0x0C, 0x02, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x64, 0x00, 0x0A, 0x00,
+                // 2 - 0x20CA220
+                0x05, 0x0A, 0x03, 0x03, 0x94, 0x7E, 0x94, 0x7E, 0x02, 0x12, 0x04, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x14, 0x14, 0x00, 0x0B, 0x0B, 0xFF, 0xFF, 0x03, 0x03, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x04, 0x00, 0x04, 0x00,
+                0x04, 0x00, 0x04, 0x00, 0x04, 0x00, 0x04, 0x00, 0x0F, 0x00, 0x0F, 0x00, 0x0F, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x90, 0x01, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00,
+                0x00, 0x20, 0x00, 0x00, 0x00, 0x40, 0x01, 0x00, 0x00, 0x40, 0x01, 0x00, 0x00, 0x40, 0x01, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x03, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x32, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00,
+                // 3 - 0x20CA040
+                0x06, 0x0A, 0x04, 0x04, 0x3F, 0x02, 0x3F, 0x02, 0x02, 0x01, 0x00, 0x40, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x02, 0x02, 0x14, 0x14, 0x00, 0x09, 0x09, 0xFF, 0xFF, 0x03, 0x03, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00, 0x05, 0x00,
+                0x05, 0x00, 0x05, 0x00, 0x05, 0x00, 0x05, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x34, 0xFF, 0xFF, 0xFF,
+                0x34, 0xFF, 0xFF, 0xFF, 0x34, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00,
+                0x00, 0x30, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x64, 0x00, 0x0A, 0x00
+            });
+        }
+
+        private static IReadOnlyList<RawWeaponInfo> Get1PWeapons()
+        {
+            return ParseWeaponInfo(count: 18, new byte[]
+            {
+                0x00, 0x00, 0x00, 0x00, 0x5F, 0x23, 0x9F, 0x53, 0x01, 0x07, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x05, 0x05, 0x00, 0x04, 0x5F, 0x41, 0x41, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x12, 0x00, 0x1E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x00, 0x06, 0x00,
+                0x24, 0x00, 0x08, 0x00, 0x08, 0x00, 0x30, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x33, 0x03, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0x0A, 0x00, 0x4B, 0x00,
+                0x01, 0x01, 0x01, 0x02, 0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x02, 0x04, 0x40, 0x00, 0x00, 0x38, 0x00,
+                0x38, 0x00, 0x02, 0x02, 0x05, 0x05, 0x00, 0x59, 0xAE, 0x3C, 0x3C, 0x00, 0x00, 0x02, 0x02, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x0E, 0x00, 0x38, 0x00,
+                0x38, 0x00, 0x15, 0x00, 0x38, 0x00, 0x38, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x07, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00, 0x00, 0x1C, 0x00, 0x00, 0x00, 0x1C, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x03, 0x00, 0x09, 0x0F, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x02, 0x02, 0x07, 0x07, 0x8C, 0x7D, 0x8C, 0x7D, 0x02, 0x02, 0x00, 0x40, 0x18, 0x00, 0x20, 0x00,
+                0x20, 0x00, 0x03, 0x03, 0x14, 0x14, 0x01, 0x08, 0xC1, 0x41, 0x41, 0x02, 0x02, 0x02, 0x02, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x2D, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x20, 0x00, 0x30, 0x00,
+                0x30, 0x00, 0x20, 0x00, 0x30, 0x00, 0x30, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00,
+                0x07, 0x00, 0x07, 0x00, 0x02, 0x00, 0x02, 0x00, 0x66, 0x06, 0x00, 0x00, 0x99, 0x09, 0x00, 0x00,
+                0x99, 0x09, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x99, 0x05, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00,
+                0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00,
+                0x00, 0x24, 0x00, 0x00, 0x00, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0xFA, 0x00, 0x19, 0x00,
+                0x03, 0x03, 0x0A, 0x0A, 0xEF, 0x3F, 0xEF, 0x3F, 0x02, 0x04, 0x00, 0x40, 0x08, 0x00, 0x08, 0x00,
+                0x08, 0x00, 0x03, 0x03, 0x0A, 0x0A, 0x00, 0xB0, 0xB0, 0x3F, 0x3F, 0x02, 0x02, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x0C, 0x00, 0x0C, 0x00,
+                0x0C, 0x00, 0x0C, 0x00, 0x0C, 0x00, 0x0C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00,
+                0x33, 0x03, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x86, 0xFF, 0xFF, 0xFF,
+                0x86, 0xFF, 0xFF, 0xFF, 0x86, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00,
+                0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x04, 0x04, 0x08, 0x08, 0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x0C, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x3C, 0x3C, 0x00, 0x1F, 0x1F, 0x42, 0x42, 0x00, 0x00, 0x03, 0x03, 0x00,
+                0x00, 0x00, 0x0A, 0x00, 0x5A, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x48, 0x00, 0x48, 0x00,
+                0x48, 0x00, 0xC8, 0x00, 0xC8, 0x00, 0xC8, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00,
+                0x00, 0x80, 0x0C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0xC8, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x05, 0x05, 0x03, 0x03, 0x94, 0x7E, 0x94, 0x7E, 0x02, 0x82, 0x00, 0x40, 0x0C, 0x00, 0x0A, 0x00,
+                0x0A, 0x00, 0x00, 0x00, 0x0F, 0x0F, 0x00, 0x0A, 0x0A, 0x3E, 0x3E, 0x03, 0x03, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x18, 0x00, 0x18, 0x00,
+                0x18, 0x00, 0x20, 0x00, 0x20, 0x00, 0x20, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x90, 0x01, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00,
+                0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x78, 0x00, 0x00, 0x00, 0x78, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0xB0, 0xDA, 0x0B, 0x02, 0xB0, 0xDA, 0x0B, 0x02,
+                0x01, 0x00, 0x03, 0x00, 0x03, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x06, 0x06, 0x04, 0x05, 0x5F, 0x3D, 0x5F, 0x3D, 0x02, 0x92, 0x02, 0x40, 0x10, 0x00, 0x1C, 0x00,
+                0x1C, 0x00, 0x03, 0x03, 0x14, 0x0F, 0x00, 0x09, 0xC2, 0x40, 0x40, 0x02, 0x02, 0x02, 0x02, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x20, 0x00, 0x38, 0x00,
+                0x38, 0x00, 0x20, 0x00, 0x38, 0x00, 0x38, 0x00, 0x2D, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00,
+                0x07, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x99, 0x05, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x86, 0xFF, 0xFF, 0xFF,
+                0x86, 0xFF, 0xFF, 0xFF, 0x86, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x07, 0x00, 0x00,
+                0x33, 0x07, 0x00, 0x00, 0x33, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x32, 0x00, 0x00, 0x00, 0x02, 0x00, 0xC8, 0x00, 0xF4, 0x01,
+                0x07, 0x07, 0x09, 0x09, 0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x04, 0x20, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xFF, 0xFF, 0x3E, 0x3E, 0x03, 0x03, 0x03, 0x03, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x2D, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x0A, 0x00, 0x0A, 0x00,
+                0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00, 0xDB, 0x0D, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00,
+                0x00, 0xF0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0xE8, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x08, 0x08, 0x0B, 0x0B, 0xFF, 0x7F, 0xFF, 0x7F, 0x05, 0x00, 0x00, 0x40, 0xC8, 0x00, 0xC8, 0x00,
+                0xC8, 0x00, 0x03, 0x03, 0x3C, 0x3C, 0x00, 0xF8, 0xF8, 0x3C, 0x3C, 0x00, 0x00, 0x03, 0x03, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x2C, 0x01, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0xC8, 0x00, 0xC8, 0x00,
+                0xC8, 0x00, 0xC8, 0x00, 0xC8, 0x00, 0xC8, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00,
+                0x07, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x99, 0x05, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00,
+                0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x90, 0x01, 0x00,
+                0x00, 0x90, 0x01, 0x00, 0x00, 0x90, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x5F, 0x23, 0x9F, 0x53, 0x01, 0x07, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x04, 0x04, 0x00, 0x04, 0x5F, 0x41, 0x41, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x12, 0x00, 0x1E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x00, 0x06, 0x00,
+                0x28, 0x00, 0x08, 0x00, 0x08, 0x00, 0x34, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x33, 0x03, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x51, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0x0A, 0x00, 0x4B, 0x00,
+                0x01, 0x01, 0x01, 0x02, 0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x02, 0x04, 0x40, 0x00, 0x00, 0x38, 0x00,
+                0x38, 0x00, 0x02, 0x02, 0x05, 0x05, 0x00, 0x59, 0x58, 0x3C, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x02, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x0E, 0x00, 0x38, 0x00,
+                0x38, 0x00, 0x15, 0x00, 0x38, 0x00, 0x38, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x07, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00, 0x00, 0x1C, 0x00, 0x00, 0x00, 0x1C, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00,
+                0x28, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x61, 0x0F, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x02, 0x02, 0x07, 0x07, 0x32, 0x7D, 0x32, 0x7D, 0x02, 0x02, 0x00, 0x40, 0x18, 0x00, 0x20, 0x00,
+                0x20, 0x00, 0x03, 0x03, 0x14, 0x14, 0x01, 0x08, 0xC1, 0x41, 0x41, 0x02, 0x02, 0x02, 0x02, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x2D, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x20, 0x00, 0x30, 0x00,
+                0x30, 0x00, 0x20, 0x00, 0x30, 0x00, 0x30, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00,
+                0x07, 0x00, 0x07, 0x00, 0x02, 0x00, 0x02, 0x00, 0x66, 0x06, 0x00, 0x00, 0x99, 0x09, 0x00, 0x00,
+                0x99, 0x09, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x99, 0x05, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00,
+                0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0C, 0x00, 0x00, 0x00, 0x0C, 0x00, 0x00, 0x00,
+                0x51, 0x00, 0x00, 0x00, 0x00, 0x20, 0x03, 0x00, 0x61, 0x0F, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00,
+                0x00, 0x24, 0x00, 0x00, 0x00, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0xFA, 0x00, 0x19, 0x00,
+                0x03, 0x03, 0x0A, 0x0A, 0xEF, 0x3F, 0xEF, 0x3F, 0x02, 0x04, 0x00, 0x40, 0x0C, 0x00, 0x0C, 0x00,
+                0x0C, 0x00, 0x03, 0x03, 0x0F, 0x0F, 0x00, 0x0E, 0x0E, 0x3F, 0x3F, 0x02, 0x02, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x12, 0x00, 0x12, 0x00,
+                0x12, 0x00, 0x12, 0x00, 0x12, 0x00, 0x12, 0x00, 0x5A, 0x00, 0x5A, 0x00, 0x5A, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5D, 0xFF, 0xFF, 0xFF,
+                0x5D, 0xFF, 0xFF, 0xFF, 0x5D, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x32, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x00, 0x00,
+                0x04, 0x04, 0x08, 0x08, 0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x0C, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x3C, 0x3C, 0x00, 0x1F, 0x1F, 0x42, 0x42, 0x00, 0x00, 0x03, 0x03, 0x00,
+                0x00, 0x00, 0x0A, 0x00, 0x5A, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x48, 0x00, 0x48, 0x00,
+                0x48, 0x00, 0xC8, 0x00, 0xC8, 0x00, 0xC8, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00,
+                0x00, 0x80, 0x0C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0xC8, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x05, 0x05, 0x03, 0x03, 0x94, 0x7E, 0x94, 0x7E, 0x02, 0x82, 0x10, 0x40, 0x0C, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x0F, 0x0F, 0x00, 0x0A, 0x00, 0x3E, 0x3E, 0x03, 0x03, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x18, 0x00, 0x0C, 0x00,
+                0x0C, 0x00, 0x20, 0x00, 0x0C, 0x00, 0x0C, 0x00, 0xFF, 0x00, 0x05, 0x00, 0x05, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x90, 0x01, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00,
+                0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x3C, 0x00, 0x00,
+                0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0x03, 0x00, 0x00, 0xC0, 0x03, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0xB0, 0xDA, 0x0B, 0x02, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x06, 0x06, 0x04, 0x05, 0x5F, 0x3D, 0x5F, 0x3D, 0x02, 0x92, 0x02, 0x40, 0x10, 0x00, 0x12, 0x00,
+                0x12, 0x00, 0x03, 0x03, 0x14, 0x14, 0x00, 0x09, 0xC3, 0x40, 0x40, 0x02, 0x02, 0x00, 0x00, 0x00,
+                0x04, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x20, 0x00, 0x30, 0x00,
+                0x30, 0x00, 0x20, 0x00, 0x30, 0x00, 0x30, 0x00, 0x2D, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00,
+                0x07, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x99, 0x05, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x86, 0xFF, 0xFF, 0xFF,
+                0x86, 0xFF, 0xFF, 0xFF, 0x86, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x07, 0x00, 0x00,
+                0x33, 0x07, 0x00, 0x00, 0x33, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x32, 0x00, 0x00, 0x00, 0x02, 0x00, 0xC8, 0x00, 0xF4, 0x01,
+                0x07, 0x07, 0x09, 0x09, 0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x04, 0x20, 0x50, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xFF, 0xFF, 0x3E, 0x3E, 0x03, 0x03, 0x03, 0x03, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x2D, 0x00, 0x0A, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x0A, 0x00, 0x0A, 0x00,
+                0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00, 0xDB, 0x0D, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00,
+                0x00, 0xF0, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0xE8, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x08, 0x08, 0x0B, 0x0B, 0xFF, 0x7F, 0xFF, 0x7F, 0x05, 0x00, 0x00, 0x40, 0x3C, 0x00, 0x3C, 0x00,
+                0x3C, 0x00, 0x03, 0x03, 0x3C, 0x3C, 0x00, 0xF8, 0xF8, 0x3C, 0x3C, 0x00, 0x00, 0x03, 0x03, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x2C, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x3C, 0x00, 0x3C, 0x00,
+                0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x2D, 0x00,
+                0x2D, 0x00, 0x2D, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00,
+                0xEB, 0x01, 0x00, 0x00, 0xEB, 0x01, 0x00, 0x00, 0xEB, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x00, 0x00, 0x00,
+                0x66, 0x00, 0x00, 0x00, 0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00,
+                0x00, 0x30, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00
+            });
+        }
+
+        private static IReadOnlyList<RawWeaponInfo> GetMPWeapons()
+        {
+            return ParseWeaponInfo(count: 18, new byte[]
+            {
+                0x00, 0x00, 0x00, 0x00,
+                0x5F, 0x23, 0x9F, 0x53, 0x01, 0x07, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x05, 0x05, 0x00, 0x04, 0x5F, 0x41, 0x41, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x00,
+                0x1E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x00, 0x06, 0x00, 0x24, 0x00, 0x08, 0x00,
+                0x08, 0x00, 0x30, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00,
+                0x00, 0x30, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0x0A, 0x00, 0x4B, 0x00, 0x01, 0x01, 0x01, 0x02,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x02, 0x04, 0x40, 0x00, 0x00, 0x38, 0x00, 0x38, 0x00, 0x02, 0x02,
+                0x05, 0x05, 0x00, 0x59, 0xAE, 0x3C, 0x3C, 0x00, 0x00, 0x02, 0x02, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x05, 0x00, 0x19, 0x00, 0x19, 0x00, 0x0E, 0x00, 0x38, 0x00, 0x38, 0x00, 0x15, 0x00,
+                0x38, 0x00, 0x38, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00, 0x07, 0x00, 0x07, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x00, 0x50, 0x00, 0x00, 0x00, 0x1C, 0x00, 0x00, 0x00, 0x1C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x20, 0x03, 0x00, 0x09, 0x0F, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x02, 0x07, 0x07,
+                0x8C, 0x7D, 0x8C, 0x7D, 0x02, 0x02, 0x00, 0x40, 0x18, 0x00, 0x20, 0x00, 0x20, 0x00, 0x03, 0x03,
+                0x14, 0x14, 0x01, 0x08, 0xC1, 0x41, 0x41, 0x02, 0x02, 0x02, 0x02, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x2D, 0x00, 0x0A, 0x00, 0x0F, 0x00, 0x0F, 0x00, 0x20, 0x00, 0x30, 0x00, 0x30, 0x00, 0x20, 0x00,
+                0x30, 0x00, 0x30, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00, 0x07, 0x00, 0x07, 0x00,
+                0x02, 0x00, 0x02, 0x00, 0x66, 0x06, 0x00, 0x00, 0x99, 0x09, 0x00, 0x00, 0x99, 0x09, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00,
+                0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x24, 0x00, 0x00,
+                0x00, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0xFA, 0x00, 0x19, 0x00, 0x03, 0x03, 0x0A, 0x0A,
+                0xEF, 0x3F, 0xEF, 0x3F, 0x02, 0x04, 0x00, 0x40, 0x08, 0x00, 0x08, 0x00, 0x08, 0x00, 0x03, 0x03,
+                0x0A, 0x0A, 0x00, 0xB0, 0xB0, 0x3F, 0x3F, 0x02, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x04, 0x00, 0x04, 0x00, 0x04, 0x00, 0x0C, 0x00, 0x0C, 0x00, 0x0C, 0x00, 0x0C, 0x00,
+                0x0C, 0x00, 0x0C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x86, 0xFF, 0xFF, 0xFF, 0x86, 0xFF, 0xFF, 0xFF,
+                0x86, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00,
+                0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x04, 0x04, 0x08, 0x08,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x0C, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x3C, 0x3C, 0x00, 0x1F, 0x1F, 0x42, 0x42, 0x00, 0x00, 0x03, 0x03, 0x00, 0x00, 0x00, 0x0A, 0x00,
+                0x5A, 0x00, 0x14, 0x00, 0x14, 0x00, 0x14, 0x00, 0x48, 0x00, 0x48, 0x00, 0x48, 0x00, 0xC8, 0x00,
+                0xC8, 0x00, 0xC8, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0xC0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0xC8, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x05, 0x03, 0x03,
+                0x94, 0x7E, 0x94, 0x7E, 0x02, 0x82, 0x00, 0x40, 0x0C, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x00, 0x00,
+                0x0F, 0x0F, 0x00, 0x0A, 0x0A, 0x3E, 0x3E, 0x03, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x05, 0x00, 0x19, 0x00, 0x19, 0x00, 0x18, 0x00, 0x18, 0x00, 0x18, 0x00, 0x20, 0x00,
+                0x20, 0x00, 0x20, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x90, 0x01, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00,
+                0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x78, 0x00, 0x00, 0x00, 0x78, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0xB0, 0xDA, 0x0B, 0x02, 0xB0, 0xDA, 0x0B, 0x02, 0x01, 0x00, 0x03, 0x00,
+                0x03, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x06, 0x04, 0x05,
+                0x5F, 0x3D, 0x5F, 0x3D, 0x02, 0x92, 0x02, 0x40, 0x10, 0x00, 0x1C, 0x00, 0x1C, 0x00, 0x03, 0x03,
+                0x14, 0x0F, 0x00, 0x09, 0xC2, 0x40, 0x40, 0x02, 0x02, 0x02, 0x02, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x0A, 0x00, 0x14, 0x00, 0x14, 0x00, 0x20, 0x00, 0x38, 0x00, 0x38, 0x00, 0x20, 0x00,
+                0x38, 0x00, 0x38, 0x00, 0x2D, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00, 0x07, 0x00, 0x07, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x33, 0x1B, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x86, 0xFF, 0xFF, 0xFF, 0x86, 0xFF, 0xFF, 0xFF,
+                0x86, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x07, 0x00, 0x00, 0x33, 0x07, 0x00, 0x00,
+                0x33, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x32, 0x00, 0x00, 0x00, 0x02, 0x00, 0xC8, 0x00, 0xF4, 0x01, 0x07, 0x07, 0x09, 0x09,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x04, 0x20, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0xFF, 0xFF, 0x3E, 0x3E, 0x03, 0x03, 0x03, 0x03, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x2D, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00,
+                0x0A, 0x00, 0x0A, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0xC0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0xF0, 0x00, 0x00, 0xDB, 0x0D, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0xE8, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x08, 0x0B, 0x0B,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x05, 0x00, 0x00, 0x40, 0xC8, 0x00, 0xC8, 0x00, 0xC8, 0x00, 0x03, 0x03,
+                0x3C, 0x3C, 0x00, 0xF8, 0xF8, 0x3C, 0x3C, 0x00, 0x00, 0x03, 0x03, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x2C, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xC8, 0x00, 0xC8, 0x00, 0xC8, 0x00, 0xC8, 0x00,
+                0xC8, 0x00, 0xC8, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00, 0x07, 0x00, 0x07, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00,
+                0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x90, 0x01, 0x00, 0x00, 0x90, 0x01, 0x00,
+                0x00, 0x90, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x5F, 0x23, 0x9F, 0x53, 0x01, 0x07, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x04, 0x04, 0x00, 0x04, 0x5F, 0x41, 0x41, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x12, 0x00,
+                0x1E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x00, 0x06, 0x00, 0x28, 0x00, 0x08, 0x00,
+                0x08, 0x00, 0x34, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00,
+                0x00, 0x30, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x51, 0x00, 0x00, 0x00,
+                0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0x0A, 0x00, 0x4B, 0x00, 0x01, 0x01, 0x01, 0x02,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x02, 0x04, 0x40, 0x00, 0x00, 0x38, 0x00, 0x38, 0x00, 0x02, 0x02,
+                0x05, 0x05, 0x00, 0x59, 0x58, 0x3C, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x05, 0x00, 0x19, 0x00, 0x19, 0x00, 0x0E, 0x00, 0x38, 0x00, 0x38, 0x00, 0x15, 0x00,
+                0x38, 0x00, 0x38, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00, 0x07, 0x00, 0x07, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x00, 0x50, 0x00, 0x00, 0x00, 0x1C, 0x00, 0x00, 0x00, 0x1C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00,
+                0x00, 0x40, 0x06, 0x00, 0x61, 0x0F, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x02, 0x07, 0x07,
+                0x32, 0x7D, 0x32, 0x7D, 0x02, 0x02, 0x00, 0x40, 0x18, 0x00, 0x20, 0x00, 0x20, 0x00, 0x03, 0x03,
+                0x14, 0x14, 0x01, 0x08, 0xC1, 0x41, 0x41, 0x02, 0x02, 0x02, 0x02, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x2D, 0x00, 0x0A, 0x00, 0x0F, 0x00, 0x0F, 0x00, 0x20, 0x00, 0x30, 0x00, 0x30, 0x00, 0x20, 0x00,
+                0x30, 0x00, 0x30, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00, 0x07, 0x00, 0x07, 0x00,
+                0x02, 0x00, 0x02, 0x00, 0x66, 0x06, 0x00, 0x00, 0x99, 0x09, 0x00, 0x00, 0x99, 0x09, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00,
+                0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x51, 0x00, 0x00, 0x00, 0x51, 0x00, 0x00, 0x00,
+                0x00, 0x20, 0x03, 0x00, 0x61, 0x0F, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x24, 0x00, 0x00,
+                0x00, 0x24, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0xFA, 0x00, 0x19, 0x00, 0x03, 0x03, 0x0A, 0x0A,
+                0xEF, 0x3F, 0xEF, 0x3F, 0x02, 0x04, 0x00, 0x40, 0x0C, 0x00, 0x0C, 0x00, 0x0C, 0x00, 0x03, 0x03,
+                0x0F, 0x0F, 0x00, 0x0E, 0x0E, 0x3F, 0x3F, 0x02, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x05, 0x00, 0x05, 0x00, 0x05, 0x00, 0x12, 0x00, 0x12, 0x00, 0x12, 0x00, 0x12, 0x00,
+                0x12, 0x00, 0x12, 0x00, 0x5A, 0x00, 0x5A, 0x00, 0x5A, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x5D, 0xFF, 0xFF, 0xFF, 0x5D, 0xFF, 0xFF, 0xFF,
+                0x5D, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x64, 0x00, 0x32, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x00, 0x00, 0x04, 0x04, 0x08, 0x08,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x0C, 0x00, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x3C, 0x3C, 0x00, 0x1F, 0x1F, 0x42, 0x42, 0x00, 0x00, 0x03, 0x03, 0x00, 0x00, 0x00, 0x0A, 0x00,
+                0x5A, 0x00, 0x14, 0x00, 0x14, 0x00, 0x14, 0x00, 0x48, 0x00, 0x48, 0x00, 0x48, 0x00, 0xC8, 0x00,
+                0xC8, 0x00, 0xC8, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0xC0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00, 0x00, 0x80, 0x0C, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0xC8, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x05, 0x03, 0x03,
+                0x94, 0x7E, 0x94, 0x7E, 0x02, 0x82, 0x10, 0x40, 0x0C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x0F, 0x0F, 0x00, 0x0A, 0x00, 0x3E, 0x3E, 0x03, 0x03, 0x00, 0x00, 0x00, 0x01, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x05, 0x00, 0x19, 0x00, 0x19, 0x00, 0x18, 0x00, 0x0C, 0x00, 0x0C, 0x00, 0x20, 0x00,
+                0x0C, 0x00, 0x0C, 0x00, 0xFF, 0x00, 0x05, 0x00, 0x05, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x20, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x90, 0x01, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00, 0x00, 0x04, 0x00, 0x00,
+                0x00, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x3C, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0x03, 0x00, 0x00, 0xC0, 0x03, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0xB0, 0xDA, 0x0B, 0x02, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x06, 0x06, 0x04, 0x05,
+                0x5F, 0x3D, 0x5F, 0x3D, 0x02, 0x92, 0x02, 0x40, 0x10, 0x00, 0x12, 0x00, 0x12, 0x00, 0x03, 0x03,
+                0x14, 0x14, 0x00, 0x09, 0xC3, 0x40, 0x40, 0x02, 0x02, 0x00, 0x00, 0x00, 0x04, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x0A, 0x00, 0x14, 0x00, 0x14, 0x00, 0x20, 0x00, 0x30, 0x00, 0x30, 0x00, 0x20, 0x00,
+                0x30, 0x00, 0x30, 0x00, 0x2D, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00, 0x07, 0x00, 0x07, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00, 0x99, 0x05, 0x00, 0x00,
+                0x33, 0x1B, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00, 0x33, 0x1B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x86, 0xFF, 0xFF, 0xFF, 0x86, 0xFF, 0xFF, 0xFF,
+                0x86, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x07, 0x00, 0x00, 0x33, 0x07, 0x00, 0x00,
+                0x33, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x32, 0x00, 0x00, 0x00, 0x02, 0x00, 0xC8, 0x00, 0xF4, 0x01, 0x07, 0x07, 0x09, 0x09,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x04, 0x20, 0x50, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0xFF, 0xFF, 0x3E, 0x3E, 0x03, 0x03, 0x03, 0x03, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x2D, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x0A, 0x00,
+                0x0A, 0x00, 0x0A, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0xC0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0xF0, 0x00, 0x00, 0xDB, 0x0D, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00, 0x00, 0xF0, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0xE8, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x08, 0x0B, 0x0B,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x05, 0x00, 0x00, 0x40, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x03, 0x03,
+                0x3C, 0x3C, 0x00, 0xF8, 0xF8, 0x3C, 0x3C, 0x00, 0x00, 0x03, 0x03, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x2C, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00,
+                0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x2D, 0x00, 0x2D, 0x00, 0x2D, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0xEB, 0x01, 0x00, 0x00,
+                0xEB, 0x01, 0x00, 0x00, 0xEB, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x00, 0x00, 0x00, 0x66, 0x00, 0x00, 0x00,
+                0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00,
+                0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00
+            });
+        }
+
+        private static IReadOnlyList<RawWeaponInfo> GetEnemyWeapons()
+        {
+            return ParseWeaponInfo(count: 11, new byte[]
+            {
+                0x00, 0x0A, 0x15, 0x15, 0x5F, 0x23, 0x9F, 0x53, 0x01, 0x07, 0x06, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x03, 0x03, 0x00, 0xF2, 0xF2, 0x41, 0x41, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x1E, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00, 0x0C, 0x00,
+                0x24, 0x00, 0x08, 0x00, 0x0C, 0x00, 0x24, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x66, 0x0A, 0x00, 0x00, 0x66, 0x0A, 0x00, 0x00, 0x66, 0x0A, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x00, 0x00, 0x00,
+                0x14, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x0A, 0x02, 0x02, 0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x02, 0xC6, 0x40, 0x14, 0x00, 0x20, 0x00,
+                0x20, 0x00, 0x00, 0x00, 0x08, 0x08, 0x00, 0x59, 0x58, 0x3C, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x02,
+                0x02, 0x00, 0x0F, 0x00, 0x2D, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x00, 0x20, 0x00,
+                0x20, 0x00, 0x1E, 0x00, 0x20, 0x00, 0x20, 0x00, 0x5A, 0x00, 0x5A, 0x00, 0x5A, 0x00, 0x07, 0x00,
+                0x07, 0x00, 0x07, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0xCC, 0x0C, 0x00, 0x00, 0xCC, 0x0C, 0x00, 0x00, 0xCC, 0x0C, 0x00, 0x00,
+                0x7A, 0x04, 0x00, 0x00, 0x7A, 0x04, 0x00, 0x00, 0x7A, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00,
+                0x00, 0x30, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x02, 0x0A, 0x07, 0x07, 0x8C, 0x7D, 0x8C, 0x7D, 0x02, 0x03, 0x06, 0x40, 0x1E, 0x00, 0x1E, 0x00,
+                0x1E, 0x00, 0x02, 0x02, 0x14, 0x14, 0x01, 0x08, 0x08, 0x41, 0x41, 0x02, 0x02, 0x02, 0x02, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x1E, 0x00, 0x1E, 0x00,
+                0x1E, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0x1E, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00,
+                0x07, 0x00, 0x07, 0x00, 0x02, 0x00, 0x02, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00,
+                0xCC, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00,
+                0x00, 0x30, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0x7D, 0x00, 0x19, 0x00,
+                0x03, 0x0A, 0x00, 0x00, 0xEF, 0x3F, 0xEF, 0x3F, 0x02, 0x04, 0x06, 0x40, 0x02, 0x00, 0x02, 0x00,
+                0x02, 0x00, 0x03, 0x03, 0x08, 0x08, 0x00, 0xB0, 0xB0, 0x3F, 0x3F, 0x03, 0x03, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0x00, 0x02, 0x00,
+                0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00, 0x00, 0x50, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x32, 0x00, 0x0A, 0x00, 0x0A, 0x00, 0x00, 0x00,
+                0x04, 0x0A, 0x08, 0x08, 0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x0C, 0x06, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x3C, 0x3C, 0x00, 0x1F, 0x1F, 0x42, 0x42, 0x03, 0x03, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x0A, 0x00, 0x5A, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x63, 0x00, 0x63, 0x00,
+                0x63, 0x00, 0xC7, 0x00, 0xC7, 0x00, 0xC7, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0x40, 0x01, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0xE0, 0x01, 0x00, 0x00, 0xE0, 0x01, 0x00, 0x00, 0xE0, 0x01, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x20, 0x03, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0xC8, 0x00, 0x00, 0x00, 0x0A, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x05, 0x0A, 0x03, 0x06, 0x94, 0x7E, 0x94, 0x7E, 0x02, 0x02, 0x16, 0x40, 0x1B, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x0F, 0x0F, 0x00, 0x0A, 0x00, 0x3E, 0x3E, 0x03, 0x03, 0x00, 0x00, 0x01,
+                0x01, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x1B, 0x00, 0x14, 0x00,
+                0x14, 0x00, 0x24, 0x00, 0x14, 0x00, 0x14, 0x00, 0x28, 0x00, 0x0F, 0x00, 0x0F, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x90, 0x01, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00,
+                0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x38, 0x00, 0x00,
+                0x00, 0x38, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xC0, 0x03, 0x00, 0x00, 0xC0, 0x03, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x20, 0xA2, 0x0C, 0x02, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x32, 0x00, 0x00, 0x00, 0x05, 0x00, 0x19, 0x00, 0x32, 0x00,
+                0x06, 0x0A, 0x16, 0x16, 0x5F, 0x3D, 0x5F, 0x3D, 0x02, 0x82, 0x06, 0x40, 0x1D, 0x00, 0x3A, 0x00,
+                0x3A, 0x00, 0x02, 0x02, 0x14, 0x14, 0x00, 0x09, 0x09, 0x40, 0x40, 0x02, 0x02, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x2D, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x1D, 0x00, 0x3A, 0x00,
+                0x3A, 0x00, 0x1D, 0x00, 0x3A, 0x00, 0x3A, 0x00, 0x1E, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x0F, 0x00,
+                0x0F, 0x00, 0x0F, 0x00, 0x02, 0x00, 0x02, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00, 0x33, 0x13, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x34, 0xFF, 0xFF, 0xFF, 0x34, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00,
+                0x00, 0x30, 0x00, 0x00, 0x00, 0x30, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x30, 0xA1, 0x0C, 0x02,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x32, 0x00, 0x00, 0x00, 0x02, 0x00, 0xC8, 0x00, 0xF4, 0x01,
+                0x07, 0x0A, 0x09, 0x09, 0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x84, 0x26, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0xFF, 0xFF, 0x3E, 0x3E, 0x03, 0x03, 0x03, 0x03, 0x00,
+                0x00, 0x00, 0x0F, 0x00, 0x2D, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x19, 0x00, 0x19, 0x00,
+                0x19, 0x00, 0x19, 0x00, 0x19, 0x00, 0x19, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00, 0x00, 0x70, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x80, 0x00, 0x00, 0x50, 0x0B, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00,
+                0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x32, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x00, 0x0A, 0x0E, 0x0E, 0x5F, 0x23, 0x9F, 0x53, 0x01, 0x07, 0x06, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x03, 0x03, 0x00, 0x04, 0x04, 0x41, 0x41, 0x03, 0x03, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x1E, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00, 0x0C, 0x00,
+                0x24, 0x00, 0x08, 0x00, 0x0C, 0x00, 0x24, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0x99, 0x09, 0x00, 0x00, 0x99, 0x09, 0x00, 0x00, 0x99, 0x09, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x00, 0x00, 0x00,
+                0x14, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x06, 0x0A, 0x0F, 0x0F, 0x5F, 0x23, 0x9F, 0x53, 0x01, 0x07, 0xC6, 0x40, 0x0F, 0x00, 0x0F, 0x00,
+                0x0F, 0x00, 0x03, 0x03, 0x03, 0x03, 0x00, 0x71, 0x71, 0x41, 0x41, 0x03, 0x03, 0x00, 0x00, 0x04,
+                0x00, 0x00, 0x1E, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00, 0x0C, 0x00,
+                0x24, 0x00, 0x08, 0x00, 0x0C, 0x00, 0x24, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00, 0x33, 0x0B, 0x00, 0x00,
+                0x33, 0x0B, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0xCC, 0x08, 0x00, 0x00, 0xCC, 0x08, 0x00, 0x00, 0xCC, 0x08, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x00, 0x00, 0x00,
+                0x14, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00,
+                0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x0A, 0x10, 0x10, 0x5F, 0x23, 0x9F, 0x53, 0x01, 0x07, 0xC6, 0x40, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x03, 0x03, 0x00, 0x86, 0x86, 0x41, 0x41, 0x03, 0x03, 0x00, 0x00, 0x01,
+                0x00, 0x00, 0x1E, 0x00, 0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00, 0x0C, 0x00,
+                0x24, 0x00, 0x08, 0x00, 0x0C, 0x00, 0x24, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00,
+                0x00, 0x08, 0x00, 0x00, 0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x99, 0x01, 0x00, 0x00, 0xCC, 0x08, 0x00, 0x00, 0xCC, 0x08, 0x00, 0x00, 0xCC, 0x08, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x00, 0x00, 0x00,
+                0x14, 0x00, 0x00, 0x00, 0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00
+            });
+        }
+
+        private static IReadOnlyList<RawWeaponInfo> GetPlatformWeapons()
+        {
+            return ParseWeaponInfo(count: 4, new byte[]
+            {
+                0x02, 0x09, 0x14, 0x14,
+                0x8C, 0x7D, 0x8C, 0x7D, 0x02, 0x03, 0x06, 0x40, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x02,
+                0x14, 0x14, 0x01, 0xBB, 0xBB, 0xBC, 0xBC, 0x02, 0x02, 0x02, 0x02, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x03, 0x00, 0x03, 0x00, 0x03, 0x00, 0x03, 0x00,
+                0x03, 0x00, 0x03, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x07, 0x00, 0x07, 0x00, 0x07, 0x00,
+                0x02, 0x00, 0x02, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x33, 0x03, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00,
+                0x00, 0x28, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x28, 0x00, 0x00, 0x00, 0xCC, 0x00, 0x00, 0x00,
+                0x00, 0x40, 0x06, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00, 0x00, 0x18, 0x00, 0x00,
+                0x00, 0x18, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x4B, 0x00, 0x19, 0x00, 0x05, 0x00, 0x7D, 0x00, 0x19, 0x00, 0x00, 0x09, 0x00, 0x00,
+                0x5F, 0x23, 0x5F, 0x23, 0x01, 0x05, 0x06, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x03, 0x03, 0x00, 0xFF, 0xFF, 0x41, 0x41, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x3C, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x05, 0x00, 0x05, 0x00, 0x05, 0x00, 0x08, 0x00,
+                0x08, 0x00, 0x08, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0xFF, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x90, 0x01, 0x00, 0x38, 0x0F, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00, 0x66, 0x0E, 0x00, 0x00,
+                0x66, 0x0E, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x64, 0x00, 0x00, 0x00, 0x01, 0x00, 0x00, 0x00, 0x00, 0x00, 0x07, 0x09, 0x11, 0x11,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x84, 0x26, 0x40, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0xFF, 0xFF, 0x3D, 0x3D, 0x03, 0x03, 0x03, 0x03, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x2D, 0x00, 0x0F, 0x00, 0x0F, 0x00, 0x0F, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00, 0x3C, 0x00,
+                0x3C, 0x00, 0x3C, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0xA0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x90, 0x03, 0x00, 0x00, 0x90, 0x03, 0x00, 0x00, 0x90, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x80, 0x00, 0x00, 0x50, 0x0B, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00, 0x00, 0x10, 0x00, 0x00,
+                0x00, 0x10, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0x32, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00, 0x01, 0x00, 0x07, 0x09, 0x09, 0x09,
+                0xFF, 0x7F, 0xFF, 0x7F, 0x02, 0x04, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0xFF, 0xFF, 0x3E, 0x3E, 0x03, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x0F, 0x00,
+                0x2D, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x14, 0x00, 0x14, 0x00, 0x14, 0x00, 0x14, 0x00,
+                0x14, 0x00, 0x14, 0x00, 0x02, 0x00, 0x02, 0x00, 0x02, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0xF0, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x20, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00, 0x99, 0x01, 0x00, 0x00,
+                0x00, 0x60, 0x00, 0x00, 0x00, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00,
+                0x66, 0x06, 0x00, 0x00, 0x66, 0x06, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00, 0xCC, 0x04, 0x00, 0x00,
+                0xCC, 0x04, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x01, 0x00, 0x01, 0x00,
+                0x01, 0x00, 0xE8, 0x03, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00
+            });
+        }
+
+        public readonly struct RawWeaponInfo
+        {
+            public readonly BeamType Weapon;
+            public readonly BeamType WeaponType; // same as Weapon except for platform/enemy beams
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 2)]
+            public readonly byte[] DrawFuncIds;
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U2, SizeConst = 2)]
+            public readonly ushort[] Colors;
+            public readonly WeaponFlags Flags;
+            public readonly ushort SplashDamage;
+            public readonly ushort MinChargeSplashDamage;
+            public readonly ushort ChargedSplashDamage;
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 2)]
+            public readonly byte[] SplashDmgTypes;
+            public readonly byte ShotCooldown;
+            public readonly byte ShotCooldownRelated;
+            public readonly byte AmmoType;
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 2)]
+            public readonly byte[] BeamTypes; // correspond to collision effects
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 2)]
+            public readonly byte[] MuzzleEffects;
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 2)]
+            public readonly byte[] DmgDirTypes;
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 2)]
+            public readonly byte[] Field1D;
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U1, SizeConst = 2)]
+            public readonly Affliction[] Afflictions; // bit 0 - freeze, bit 1 - disrupt, bit 3 - burn
+            public readonly byte Field21;
+            public readonly ushort MinCharge;
+            public readonly ushort FullCharge;
+            public readonly ushort AmmoCost;
+            public readonly ushort MinChargeCost;
+            public readonly ushort ChargeCost;
+            public readonly ushort UnchargedDamage;
+            public readonly ushort MinChargeDamage;
+            public readonly ushort ChargedDamage;
+            public readonly ushort HeadshotDamage;
+            public readonly ushort MinChargeHeadshotDamage;
+            public readonly ushort ChargedHeadshotDamage;
+            public readonly ushort UnchargedLifespan;
+            public readonly ushort MinChargeLifespan;
+            public readonly ushort ChargedLifespan;
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U2, SizeConst = 2)]
+            public readonly ushort[] SpeedDecay;
+            public readonly ushort Field42;
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U2, SizeConst = 2)]
+            public readonly ushort[] SpeedInterp;
+            public readonly int Field48; // uncharged
+            public readonly int Field4C; // min charge
+            public readonly int Field50; // full charge
+            public readonly int Field54;
+            public readonly int Field58; // uncharged
+            public readonly int Field5C; // min charge
+            public readonly int Field60; // full charge
+            public readonly int UnchargedSpeed; // uncharged
+            public readonly int MinChargeSpeed; // min charge
+            public readonly int ChargedSpeed; // full charge
+            public readonly int UnchargedFinalSpeed; // uncharged
+            public readonly int MinChargeFinalSpeed; // min charge
+            public readonly int ChargedFinalSpeed; // full charge
+            public readonly int UnchargedGravity;
+            public readonly int MinChargeGravity;
+            public readonly int ChargedGravity;
+            public readonly int UnchargedHoming;
+            public readonly int MinChargeHoming;
+            public readonly int ChargedHoming;
+            public readonly int HomingRange;
+            public readonly int HomingTolerance;
+            public readonly int UnchargedScale;
+            public readonly int MinChargeScale;
+            public readonly int ChargedScale;
+            public readonly int UnchargedDistance;
+            public readonly int MinChargeDistance;
+            public readonly int ChargedDistance;
+            public readonly int UnchargedSpread; // uncharged
+            public readonly int MinChargeSpread; // min charge
+            public readonly int ChargedSpread; // full charge
+            public readonly int FieldC0; // uncharged
+            public readonly int FieldC4; // min charge
+            public readonly int FieldC8; // full charge
+            public readonly int FieldCC; // uncharged
+            public readonly int FieldD0; // min charge
+            public readonly int FieldD4; // full charge
+            [MarshalAs(UnmanagedType.ByValArray, ArraySubType = UnmanagedType.U4, SizeConst = 2)]
+            public readonly uint[] RicochetWeaponPtr; // WeaponInfo*
+            public readonly ushort ProjectileCount;
+            public readonly ushort MinChargedProjectileCount;
+            public readonly ushort ChargeProjectileCount;
+            public readonly ushort SmokeStart; // start drawing gun smoke when smoke level reaches this value (and cap it)
+            public readonly ushort SmokeMinimum; // continue drawing gun smoke until level drops below this value
+            public readonly ushort SmokeDrain; // reduce level by this amount each frame (or bring it to 0)
+            public readonly ushort SmokeShotAmount; // increase level by this amount when firing uncharged shot
+            public readonly ushort SmokeChargeAmount; // increase level by this amount each frame while charging
+        }
+
+        private static void Nop()
+        {
+        }
+    }
+}
